framework for per-priority mixing

+++ app/src/main/java/com/qfs/apres/soundfontplayer/FrameMap.kt
@@ -2,7 +2,7 @@ package com.qfs.apres.soundfontplayer
 
 interface FrameMap {
     fun get_new_handles(frame: Int): List<Set<SampleHandle>>?
-    fun get_track_priority(track: Int): Float
+    fun get_track_priority(track: Int): Int
     fun get_beat_frames(): HashMap<Int, IntRange>
     fun get_active_handles(frame: Int): List<Set<Pair<Int, SampleHandle>>>
     fun get_size(): Int
